Foundational code for gathering text from CapitolWordsAPI
MessageObject is the text of the legislator speaking.
CandidateObject is the legislator, which has a list of MessageObjects.
main - currently accepts keyword/phrase and searches for keyword in text
of legislators who speak. It takes date, title, messages, first name,
last name, etc. that will be used later for runtime manipulation.